api/requirements.txt
Django==5.1.2
djangorestframework
psycopg2-binary
python-dateutil
pytest
pytest-django
pytest-mockito
factory-boy
injector
pymemcache
djangorestframework-simplejwt
django-ratelimit
drf-spectacular
phonenumbers
django-environ
django-crontab 
twilio
stripe